+++ src/Overloaded/Numerals.hs
@@ -16,6 +16,8 @@ module Overloaded.Numerals (
     defaultFromNumeral,
     ) where
 
+import Data.Bin.Pos    (Pos (..))
+import Data.BinP.PosP  (PosP (..))
 import Data.Fin        (Fin (..))
 import Data.Proxy      (Proxy (..))
 import Data.Word       (Word16, Word32, Word64, Word8)
@@ -24,7 +26,11 @@ import GHC.TypeLits    (ErrorMessage (..), Symbol, TypeError)
 import GHC.TypeNats    (type (<=?), KnownNat, Nat, natVal)
 import Numeric.Natural (Natural)
 
-import qualified Data.Type.Nat as N
+import qualified Data.Bin       as B
+import qualified Data.BinP.PosP as PP
+import qualified Data.Type.Bin  as B
+import qualified Data.Type.BinP as BP
+import qualified Data.Type.Nat  as N
 
 -- | Another way to desugar numerals.
 --
@@ -99,10 +105,12 @@ instance (KnownNat n, OverflowCheck n "Word8" (n <=? 0xffffffffffffffff)) => Fro
 instance KnownNat n => FromNumeral n N.Nat where
     fromNumeral = defaultFromNumeral @n
 
-type family IsLess (n :: N.Nat) (m :: N.Nat) (p :: Nat) (q :: Nat) :: Constraint where
-    IsLess 'N.Z     ('N.S m) p q = ()
-    IsLess ('N.S n) ('N.S m) p q = IsLess n m p q
-    IsLess ('N.S n) 'N.Z     p q = TypeError ('ShowType p ':<>: 'Text " is not less than " ':<>: 'ShowType q)
+type family IsLess (p :: Nat) (q :: Nat) :: Constraint where
+    IsLess p q = IsLess' (q <=? p) p q
+
+type family IsLess' (b :: Bool) (p :: Nat) (q :: Nat) :: Constraint where
+    IsLess' 'False p q = ()
+    IsLess' 'True  p q = TypeError ('ShowType p ':<>: 'Text " is not less than " ':<>: 'ShowType q)
 
 class FinFromNumeral (n :: N.Nat) (m :: N.Nat) where
     finFromNumeral :: Proxy n -> Fin m
@@ -113,5 +121,30 @@ instance FinFromNumeral 'N.Z ('N.S m) where
 instance FinFromNumeral n m => FinFromNumeral ('N.S n) ('N.S m) where
     finFromNumeral _ = FS (finFromNumeral (Proxy :: Proxy n))
 
-instance (FinFromNumeral (N.FromGHC n) m, IsLess (N.FromGHC n) m n (N.ToGHC m)) => FromNumeral n (Fin m) where
+instance (FinFromNumeral (N.FromGHC n) m, IsLess n (N.ToGHC m)) => FromNumeral n (Fin m) where
     fromNumeral = finFromNumeral (Proxy :: Proxy (N.FromGHC n))
+
+-------------------------------------------------------------------------------
+-- bin
+-------------------------------------------------------------------------------
+
+instance KnownNat n => FromNumeral n B.Bin where
+    fromNumeral = defaultFromNumeral @n
+
+instance (KnownNat n, (1 <=? n) ~ 'True) => FromNumeral n B.BinP where
+    fromNumeral = defaultFromNumeral @n
+
+class PosFromNumeral (n :: N.Nat) (b :: B.BinP) where
+    posFromNumeral:: Proxy n -> PosP b
+
+instance B.SBinPI b => PosFromNumeral 'N.Z b where
+    posFromNumeral _ = PP.top
+
+instance (B.SBinPI bp, PosFromNumeral n bp, B.Pred b ~ 'B.BP bp, BP.Succ bp ~ b) => PosFromNumeral ('N.S n) b where
+    posFromNumeral _ = PP.pop (posFromNumeral (Proxy :: Proxy n))
+
+instance (PosFromNumeral (N.FromGHC n) b, IsLess n (BP.ToGHC b)) => FromNumeral n (PosP b) where
+    fromNumeral = posFromNumeral (Proxy :: Proxy (N.FromGHC n))
+
+instance (PosFromNumeral (N.FromGHC n) b, IsLess n (BP.ToGHC b)) => FromNumeral n (Pos ('B.BP b)) where
+    fromNumeral = Pos (posFromNumeral (Proxy :: Proxy (N.FromGHC n)))