+++ src/Overloaded/TypeNats.hs
@@ -6,7 +6,11 @@ module Overloaded.TypeNats where
 
 import GHC.TypeNats (Nat)
 
-import qualified Data.Type.Nat as N
+import qualified Data.Bin       as B
+import qualified Data.BinP      as BP
+import qualified Data.Type.Bin  as B
+import qualified Data.Type.BinP as BP
+import qualified Data.Type.Nat  as N
 
 -- | A way to overload type level 'Nat's.
 --
@@ -30,3 +34,9 @@ instance FromNatC Nat where
 
 instance FromNatC N.Nat where
     type FromNat n = N.FromGHC n
+
+instance FromNatC B.Bin where
+    type FromNat n = B.FromGHC n
+
+instance FromNatC BP.BinP where
+    type FromNat n = BP.FromGHC n