+++ test/Overloaded/Test/Numerals.hs
@@ -2,10 +2,14 @@
 {-# OPTIONS -fplugin=Overloaded -fplugin-opt=Overloaded:Numerals #-}
 module Overloaded.Test.Numerals where
 
+import Data.Bin         (Bin, BinP)
 import Data.Fin         (Fin (..))
 import Test.Tasty       (TestTree, testGroup)
 import Test.Tasty.HUnit (testCase, (@?=))
 
+import qualified Data.Bin      as B
+import qualified Data.BinP     as BP
+import qualified Data.Type.Bin as B
 import qualified Data.Type.Nat as N
 
 tests :: TestTree
@@ -21,4 +25,16 @@ tests = testGroup "Numerals"
                 2 -> True -- works, because there is @Num (Fin n)@ instance
                 _ -> False
         res @?= True
+
+    , testCase "Data.Bin" $ do
+        let b :: Bin
+            b = 0
+
+        b @?= B.fromNatural 0
+
+    , testCase "Data.BinP" $ do
+        let b :: BinP
+            b = 1
+
+        b @?= BP.fromNatural 1
     ]