+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Make Vec, NP and POP instances match more eagerly
 - Add `Overloaded.Lists.Bidi`
 - Add `Seq`, `Map` and `IntMap` `Cons` and `Nil` instances
+- Add `bin` and `ral` instances
 
 # 0.1.3